+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file. This projec
 * Added Status Messages to logs for discoverDevices request.
 * Added Cached Status to IR device, Status will be saved to accessory context and restored on restart.
 * Added Option `Offline as Off` to be able set the device as off, if API reports offline.
+* Removed Meter Unit Config Option as it was confusing and probably never used.
 * Housekeeping and updated dependencies.
 
 **Full Changelog**: https://github.com/OpenWonderLabs/homebridge-switchbot/compare/v1.3.0...v1.3.1